Cleanup Ahrs Code, Add Reset
